clk: zynqmp: Add support for custom type flags
Store extra custom type flags received from firmware.

+++ drivers/clk/zynqmp/divider.c
@@ -36,7 +36,7 @@
  */
 struct zynqmp_clk_divider {
 	struct clk_hw hw;
-	u8 flags;
+	u16 flags;
 	u32 clk_id;
 	u32 div_type;
 };
